ShortsYT/requirements.txt
# Core
python-dotenv
pydantic

# YouTube API & Research
google-api-python-client
pandas

# AI/LLM (Script Generation)
openai-agents
litellm

# Text-to-Speech
elevenlabs

# Video Processing
moviepy
